+++ consensus/clique/src/main/java/org/hyperledger/besu/consensus/clique/CliqueHelpers.java
@@ -36,6 +36,16 @@ public class CliqueHelpers {
     return proposerSelector.selectProposerForNextBlock(parent);
   }
 
+  public static boolean isSigner(
+      final Address candidate, final ProtocolContext protocolContext, final BlockHeader parent) {
+    final VoteTally validatorProvider =
+        protocolContext
+            .getConsensusState(CliqueContext.class)
+            .getVoteTallyCache()
+            .getVoteTallyAfterBlock(parent);
+    return validatorProvider.getValidators().contains(candidate);
+  }
+
   public static boolean addressIsAllowedToProduceNextBlock(
       final Address candidate, final ProtocolContext protocolContext, final BlockHeader parent) {
     final VoteTally validatorProvider =
@@ -44,7 +54,7 @@ public class CliqueHelpers {
             .getVoteTallyCache()
             .getVoteTallyAfterBlock(parent);
 
-    if (!validatorProvider.getValidators().contains(candidate)) {
+    if (!isSigner(candidate, protocolContext, parent)) {
       return false;
     }
